# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personnel Monitoring System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /personnel:
    get:
      operationId: listPersonnel
      tags: [personnel]
      summary: List all personnel
      responses:
        "200":
          description: List of personnel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Personnel"
    post:
      operationId: createPersonnel
      tags: [personnel]
      summary: Create personnel and optionally an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonnelRequest"
      responses:
        "201":
          description: Personnel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /personnel/{id}:
    get:
      operationId: getPersonnel
      tags: [personnel]
      summary: Get a personnel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Personnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePersonnel
      tags: [personnel]
      summary: Update personnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonnelRequest"
      responses:
        "200":
          description: Updated personnel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePersonnel
      tags: [personnel]
      summary: Delete personnel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create account for existing personnel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Validation error or duplicate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Personnel:
      type: object
      properties:
        id:
          type: integer
        lastName:
          type: string
        firstName:
          type: string
        middleInitial:
          type: string
          nullable: true
        employeeId:
          type: string
        department:
          type: string
        position:
          type: string
        photoUrl:
          type: string
          nullable: true
        hasAccount:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - lastName
        - firstName
        - employeeId
        - department
        - position
        - hasAccount
        - createdAt
    CreatePersonnelRequest:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        middleInitial:
          type: string
          nullable: true
        employeeId:
          type: string
        department:
          type: string
        position:
          type: string
        photoUrl:
          type: string
          nullable: true
        createAccount:
          type: boolean
        password:
          type: string
          nullable: true
      required:
        - lastName
        - firstName
        - employeeId
        - department
        - position
    UpdatePersonnelRequest:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        middleInitial:
          type: string
          nullable: true
        department:
          type: string
        position:
          type: string
        photoUrl:
          type: string
          nullable: true
    Account:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        personnelId:
          type: integer
          nullable: true
        personnelName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - username
        - role
        - createdAt
    CreateAccountRequest:
      type: object
      properties:
        personnelId:
          type: integer
        password:
          type: string
      required:
        - personnelId
        - password
...